XF3-5980 : Moca isolation service in failed state
Reason for change: with RDKB-32485 changes,
we are killing xacl-device,xdiscovery but in
case of feature not enabled, this kill commands
won't succeed, resulting in return of error to systemd
Test Procedure: Test moca isolation service not in failed
state
Risks: Low

(cherry picked from commit 7bda92f34c5e4476157c54dbddb603fb83a30f2d)

+++ scripts/MoCA_isolation.sh
@@ -71,6 +71,12 @@ if [ "$BOX_TYPE" = "XB3" ]; then
    sh /lib/rdk/start-upnp-service restart > /dev/null  2>&1 &
 else
    # for other devices killing the process is enough, services would restart.
-   killall xcal-device
-   killall xdiscovery
+   pid_xcal_device=`pidof xcal-device`
+   if [ "x$pid_xcal_device" != "x" ];then
+   	killall xcal-device
+   fi
+   pid_xdiscovery=`pidof xdiscovery`
+   if [ "x$pid_xdiscovery" != "x" ];then
+   	killall xdiscovery
+   fi
 fi
